move scope details to own file
